requirements.txt
crewai
langchain-google-genai
langchain-community
tavily-python
streamlit
python-dotenv